refactor(backend): remove redundant HITL approval gathering and dead submitApproval signal
- Rename interrupt_proxy.py to subagent.py; update 26 patch targets
- Delete Python-side pending_approvals computation (build_snapshot_from_interrupts,
  build_pending_approvals_snapshot, slim_status_for_temporal snapshot copy)
- Go workflow reads pendingCount from DB via loadExecution() instead of slim
  activity return, matching the Java pattern
- Delete dead SignalApproval() method and SignalSubmitApproval constant
- Extract StatusBuilder handler modules (sub_agent, context, formatting,
  streaming_buffers) — follow-on from ExecutionState extraction
- Clean stale InterruptProxy/submitApproval comment references
- ~2,300 net lines deleted

Made-with: Cursor
