+++ backend/libs/python/graphton/tests/core/test_subagent_guardrails.py
@@ -18,7 +18,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from graphton.core.execution_budget import ExecutionBudgetMiddleware
-from graphton.core.interrupt_proxy import (
+from graphton.core.subagent import (
     _SUB_AGENT_ADVISORY_INTERVAL,
     _SUB_AGENT_MAX_ADVISORIES,
     _UNLIMITED_RECURSION,
@@ -46,7 +46,7 @@ def mock_tools():
 class TestGuardrailInjection:
     """Verify that compile_subagent injects guardrail middleware."""
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_injects_loop_detection(self, mock_create_agent, mock_model, mock_tools):
         mock_graph = MagicMock()
         mock_graph.with_config = MagicMock(return_value=mock_graph)
@@ -66,7 +66,7 @@ class TestGuardrailInjection:
         loop_detection = [m for m in middleware_list if isinstance(m, LoopDetectionMiddleware)]
         assert len(loop_detection) == 1
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_injects_tool_truncation(self, mock_create_agent, mock_model, mock_tools):
         mock_graph = MagicMock()
         mock_graph.with_config = MagicMock(return_value=mock_graph)
@@ -86,7 +86,7 @@ class TestGuardrailInjection:
         truncation_mw = [m for m in middleware_list if isinstance(m, ToolTruncationMiddleware)]
         assert len(truncation_mw) == 1
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_injects_execution_budget_periodic(self, mock_create_agent, mock_model, mock_tools):
         """ExecutionBudgetMiddleware is injected in periodic mode."""
         mock_graph = MagicMock()
@@ -112,7 +112,7 @@ class TestGuardrailInjection:
         assert mw.warning_interval == _SUB_AGENT_ADVISORY_INTERVAL
         assert mw.max_warnings == _SUB_AGENT_MAX_ADVISORIES
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_preserves_caller_middleware(self, mock_create_agent, mock_model, mock_tools):
         """Caller-provided middleware is preserved alongside auto-injected ones."""
         mock_graph = MagicMock()
@@ -137,7 +137,7 @@ class TestGuardrailInjection:
         assert sentinel in middleware_list
         assert len(middleware_list) == 4  # sentinel + 3 guardrails
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_total_guardrail_count(self, mock_create_agent, mock_model, mock_tools):
         """Three guardrails are injected: loop detection, truncation, budget."""
         mock_graph = MagicMock()
@@ -157,7 +157,7 @@ class TestGuardrailInjection:
 
         assert len(middleware_list) == 3
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_no_checkpointer_passed(self, mock_create_agent, mock_model, mock_tools):
         """compile_subagent must NOT pass checkpointer to create_agent
         (LangGraph per-invocation mode: sub-agent inherits parent's)."""
@@ -176,7 +176,7 @@ class TestGuardrailInjection:
         call_kwargs = mock_create_agent.call_args
         assert "checkpointer" not in (call_kwargs.kwargs or {})
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_returns_compiled_graph_directly(self, mock_create_agent, mock_model, mock_tools):
         """The returned runnable should be the compiled graph, not a proxy wrapper."""
         mock_graph = MagicMock()
@@ -200,7 +200,7 @@ class TestGuardrailInjection:
 class TestRecursionLimitForwarding:
     """Verify recursion_limit is forwarded to the compiled graph."""
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_default_recursion_limit_is_unlimited(self, mock_create_agent, mock_model, mock_tools):
         mock_graph = MagicMock()
         mock_graph.with_config = MagicMock(return_value=mock_graph)
@@ -218,7 +218,7 @@ class TestRecursionLimitForwarding:
             {"recursion_limit": _UNLIMITED_RECURSION},
         )
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_custom_recursion_limit(self, mock_create_agent, mock_model, mock_tools):
         mock_graph = MagicMock()
         mock_graph.with_config = MagicMock(return_value=mock_graph)
@@ -237,7 +237,7 @@ class TestRecursionLimitForwarding:
             {"recursion_limit": 100},
         )
 
-    @patch("graphton.core.interrupt_proxy.create_agent")
+    @patch("graphton.core.subagent.create_agent")
     def test_none_recursion_limit_uses_unlimited(self, mock_create_agent, mock_model, mock_tools):
         """Passing recursion_limit=None falls back to unlimited."""
         mock_graph = MagicMock()
