+++ backend/services/stigmer-server/pkg/domain/agentexecution/temporal/workflow_types.go
@@ -10,11 +10,6 @@ const (
 	// Handles both workflows and Python activities on the same queue.
 	AgentExecutionTaskQueue = "execution"
 
-	// SignalSubmitApproval is the signal name for submitting HITL approval decisions.
-	// This signal is sent to a running workflow when a user approves/skips/rejects a tool call.
-	// The workflow receives this signal and unblocks its Workflow.await() to resume execution.
-	SignalSubmitApproval = "submitApproval"
-
 	// SignalApprovalGateResolved is sent when the approval gate for a HITL cycle
 	// has fully resolved: either all pending tool calls have decisions, or a REJECT
 	// was submitted (which triggers immediate resume — Python auto-skips remaining
